added isAllowed property to each applicable endpoint

+++ src/routers/list-router.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const ListService = require('../services/list-service');
+const UserService = require('../services/user-service');
 const Security = require('../helpers/security')
 
 const ListRouter = express.Router();
@@ -7,6 +8,10 @@ const ListRouter = express.Router();
 ListRouter.route('/')
   .get(async (req, res, next) => {
     const db = req.app.get('db');
+
+    if (!req.admin) {
+      return next('Unauthorized access');
+    };
 
     const lists = await ListService.getAll(db)
       .catch(next);
@@ -17,11 +22,19 @@ ListRouter.route('/')
     const db = req.app.get('db');
     const { name, user_id } = req.body;
 
+    if (user_id !== req.user_id && !req.admin) {
+      return next('Unauthorized access');
+    };
+
     if (!name) next('Invalid data.');
 
     const list = Security.applyXSS({ name });
 
     if (user_id) {
+      await UserService.findByID(user_id)
+        .catch(next);
+      // Check if user exists
+
       list.user_id = user_id;
     };
 
@@ -49,11 +62,22 @@ ListRouter.route('/:list')
     const list = await ListService.findByID(db, list_id)
       .catch(next);
     
+    if (list.user_id !== req.user_id && !req.admin) {
+      return next('Unauthorized access');
+    };  
+    
     return res.json(list);
   })
   .patch(async (req, res, next) => {
     const db = req.app.get('db');
     const list_id = parseInt(req.params.list);
+
+    const list = await ListService.findByID(db, list_id)
+      .catch(next);
+
+    if (list.user_id !== req.user_id && !req.admin) {
+      return next('Unauthorized access');
+    };  
 
     const { name } = req.body;
     const newValues = Security.applyXSS({ name });
@@ -66,6 +90,13 @@ ListRouter.route('/:list')
   .delete(async (req, res, next) => {
     const db = req.app.get('db');
     const list_id = parseInt(req.params.list);
+
+    const list = await ListService.findByID(db, list_id)
+      .catch(next);
+
+    if (list.user_id !== req.user_id && !req.admin) {
+      return next('Unauthorized access');
+    };
 
     await ListService.delete(db, list_id)
       .catch(next);
